+++ arrakis/lib/options/options.ml
@@ -17,7 +17,7 @@ let unix_file   = ref (Unix.getcwd () ^ "/socket")
 
 let no_color   = ref false
 let allow_root = ref false
-let no_shell   = ref false
+let just_run   = ref false
 
 let env = ref "unix"
 
@@ -27,8 +27,8 @@ let spec = [
   ("-e",            Arg.Set_string env,        "<venus|unix> Set env for ecalls"  );
   ("--no-color",    Arg.Set no_color,          " Don't use color in output"       );
   ("--allow-root",  Arg.Set allow_root,        " Allow usage in root mode"        );
-  ("-ns",           Arg.Set no_shell,          " Run the program and exit"        );
-  ("--no-shell",    Arg.Set no_shell,          " Run the program and exit"        );
+  ("-r",            Arg.Set just_run,          " Run the program and exit"        );
+  ("--just-run",    Arg.Set just_run,          " Run the program and exit"        );
   ("--version",     Arg.Set show_version,      " Show version number and exit"    );
 ]
 
@@ -46,6 +46,6 @@ let unix_file   = !unix_file
 
 let no_color    = !no_color
 let allow_root  = !allow_root
-let no_shell    = !no_shell
+let just_run    = !just_run
 
 let env = !env